pyproject.toml
[project]
name = "env-cli"
version = "0.1.0"
description = "env-cli"
authors = [
    {name = "Raanan Denis", email = "raanand@gmail.com"},
]
dependencies = [
    "click>=8.3.1",
    "boto3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "pygithub>=2.8.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]

env-cli = "env_cli.main:cli"


